Implement forward declaration support
It's not too pretty unfortunately - there are now two function instances, and
one forwards to another. Not sure if there's a better way to do this - we can
split HLSLFunction declaration and definition nodes but we need to link
FunctionCall instances to something, and to keep forward declaration and the
definition nodes visible if we want to translate all of them.

Alternatively we could try to just automatically insert forward declaration
nodes in the output...