GLSL: Add constant buffer emulation for vec4 arrays
Normally we don't support arrays because array literals are not supported by
legacy GLSL, but if there's just one vec4 array it's easy to just return it.

At some point we might implement compile-time evaluation of field indexing or
array access, but it's much easier to use a struct getter for now.