MSL: Various generation improvements
Stop using FindFunctionCall - it does not work properly because it does not
guard function definition emission. The correct approach is to use
NeedsFunctionCall, and to pre-prune a tree if necessary.

Always emit texture structures - they don't hurt anything, and without
these it's impossible to have a function that accepts a sampler and does not
use it...

Also add float4x4 -> float3x3 casting support.