GLSL: Introduce Options struct
Move flags into Options and introduce constant buffer prefix.

By default we use the name from the HLSL source for the constant buffer name in
both emulated and non-emulated paths. In some cases it's important to customize
the buffer name - for example, on some Adreno drivers on Android GLSL linker
can't correctly merge the array sizes and the only way to avoid these bugs is
to have differently named uniform arrays for VS & FS. This option allows for
precisely that, and possibly other use cases.