MSL: Fix 'static const' variables
Static const variables are not supported in function scope (because of Metal
restrictions), and also aren't supported in class scope (C++ linking rules).

For them to work, we need to lift them all the way up to global scope because
of this.

This lifting is somewhat naive - it doesn't resolve naming conflicts and does
not lift struct declarations - but it's better than nothing!